+++ client/src/pages/market/sources.jsx
@@ -31,19 +31,19 @@ import { CosmosCheckbox, CosmosFormDivider, CosmosInputText, CosmosSelect } from
 import ResponsiveButton from '../../components/responseiveButton';
 import { useTranslation } from 'react-i18next';
 
-const AlertValidationSchema = Yup.object().shape({
-  name: Yup.string().required('Name is required'),
-  trackingMetric: Yup.string().required('Tracking metric is required'),
-  conditionOperator: Yup.string().required('Condition operator is required'),
-  conditionValue: Yup.number().required('Condition value is required'),
-  period: Yup.string().required('Period is required'),
-});
-
 const EditSourcesModal = ({ onSave }) => {
   const { t } = useTranslation();
   const [config, setConfig] = React.useState(null);
   const [open, setOpen] = React.useState(false);
   
+  const AlertValidationSchema = Yup.object().shape({
+    name: Yup.string().required(t('global.name.validation')),
+    trackingMetric: Yup.string().required(t('navigation.monitoring.alerts.action.edit.trackingMetric.validation')),
+    conditionOperator: Yup.string().required(t('navigation.monitoring.alerts.action.edit.conditionOperator.validation')),
+    conditionValue: Yup.number().required(t('navigation.monitoring.alerts.action.edit.conditionValue.validation')),
+    period: Yup.string().required(t('navigation.monitoring.alerts.action.edit.period.validation')),
+  });
+
   function getConfig() {
     API.config.get().then((res) => {
       setConfig(res.data);